fix(deps): migrate MCP route to mcp-handler v2
mcp-handler 2.x drops the 3rd config arg (basePath/maxDuration) and the SSE transport, and swaps its required peer dep to @modelcontextprotocol/server v2. Move the handler to a static /api/mcp route (the [transport] dynamic segment only existed for the now-removed SSE transport), collapse createMcpHandler to its one-arg form, and re-express maxDuration as a Next.js route-segment export. Drop the orphaned @modelcontextprotocol/sdk v1 (imported nowhere after this). Verified: next build, 87/87 tests, plus a live initialize + tools/list against the built /api/mcp endpoint.
