+++ package.json
@@ -20,7 +20,7 @@
     "@base-ui/react": "^1.6.0",
     "@codemirror/lang-markdown": "^6.5.1",
     "@codemirror/view": "^6.43.6",
-    "@modelcontextprotocol/sdk": "^1.26.0",
+    "@modelcontextprotocol/server": "^2.0.0",
     "@next/third-parties": "^16.2.10",
     "@radix-ui/react-dialog": "^1.1.20",
     "@radix-ui/react-dropdown-menu": "^2.1.21",
@@ -38,7 +38,7 @@
     "framer-motion": "^12.42.2",
     "github-slugger": "^2.0.0",
     "lucide-react": "^1.25.0",
-    "mcp-handler": "^1.1.0",
+    "mcp-handler": "^2.1.1",
     "mermaid": "^11.16.0",
     "next": "16.3.1",
     "next-themes": "^0.4.6",
